Fix some warnings. allow gnu license (max exporter)
git-svn-id: svn+ssh://svn.corp.imvu.com/home/svnrepos/trunk/stable/Source/cal3d@114339 07c76cb3-cb09-0410-85de-c24d39f1912e

+++ plugins/src/SubmeshCandidate.cpp
@@ -136,7 +136,7 @@ bool CSubmeshCandidate::AddSpring(int vertexId1, int vertexId2)
 int CSubmeshCandidate::AddVertexCandidate(CVertexCandidate *pVertexCandidate)
 {
 	// check if the vertex candidate has the same number of maps as this submesh candidate
-	if(pVertexCandidate->GetVectorTextureCoordinate().size() != m_mapCount)
+	if(pVertexCandidate->GetVectorTextureCoordinate().size() != unsigned(m_mapCount))
 	{
 		theExporter.SetLastError("Map count mismatch (vertex <-> submesh)!", __FILE__, __LINE__);
 		return -1;
